Adjusted a contrib plugins inline hyperlink
Originally referenced the the same URL found near the end of line 113. This line was " Check out our link:https://community.taiga.io/t/how-can-i-contribute/159[detailed contribution guide] if you'd like to support Taiga. " Note: The end of the url is " how-can-i-contribute/159"

This was found on the line I edited (118)

I believe it is hyperlinked wrong as it references a "You can check a list of contributions." but then links to the "how can I contribute."

I hunted down the URL and believe the one I have edited in is the RIGHT one.